requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
Django==3.2.4
django-bootstrap-calendar==0.1.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==3.0.1
django-calendarium==1.3.4
django-filer==2.0.2
django-js-asset==1.2.2
django-libs==2.0.3
django-mptt==0.13.0
django-phonenumber-field==5.2.0
django-polymorphic==3.0.0
djangorestframework==3.12.4
easy-thumbnails==2.7.1
jsmin==2.2.2
phonenumbers==8.12.28
phonenumberslite==8.12.28
Pillow==8.3.1
psycopg2==2.9.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
Unidecode==1.1.2
